v4 done

+++ src/main/java/com/ccor/ecommerce/model/dto/ProductSoldResponseDTO.java
@@ -5,6 +5,6 @@ public record ProductSoldResponseDTO(
         String barCode,
         String name,
         int amount,
-        double price
+        int price
 ) {
 }
